+++ Dashboard/SimuMan/servergite.cpp
@@ -39,12 +39,12 @@ void servergite::bytesWritten(qint64 nb){
      qDebug() << "" << nb << " Bytes sent ..." ;
 }
 
-void servergite::update(int newval)
+void servergite::update(float newval)
 {
 
     if (isconnected==true){
 
-        float f = (float)newval;
+        float f = newval;
          //qDebug() <<f;
         QByteArray x(reinterpret_cast<const char *>(&f), sizeof(f)) ;
         a=x;
